update backend (AddPlayer api2)

+++ ttf-backend/src/main/java/io/aleksb/springjwt/ttfmain/repository/GlobalRepository.java
@@ -3,5 +3,5 @@ package io.aleksb.springjwt.ttfmain.repository;
 import io.aleksb.springjwt.ttfmain.models.Global;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GlobalRepository extends JpaRepository<Global,Long> {
+public interface GlobalRepository extends JpaRepository<Global,Integer> {
 }
